Move `StructuredOutputsConfig` from `config/__init__.py` to `config/structured_outputs.py`
